# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skymark Eatery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  displayOrder: number;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  /** @nullable */
  description?: string | null;
  displayOrder: number;
}

export interface UpdateCategoryBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  displayOrder?: number | null;
}

export interface MenuItem {
  id: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  available: boolean;
  isPopular: boolean;
  isFeatured?: boolean;
  /** @nullable */
  allergens?: string | null;
  createdAt: string;
}

export interface CreateMenuItemBody {
  /** @nullable */
  categoryId?: number | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  available?: boolean;
  isPopular?: boolean;
  isFeatured?: boolean;
  /** @nullable */
  allergens?: string | null;
}

export interface UpdateMenuItemBody {
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  available?: boolean | null;
  /** @nullable */
  isPopular?: boolean | null;
  /** @nullable */
  isFeatured?: boolean | null;
  /** @nullable */
  allergens?: string | null;
}

export interface Special {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  activeDate: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateSpecialBody {
  title: string;
  description: string;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  activeDate: string;
  isActive?: boolean;
}

export interface UpdateSpecialBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  activeDate?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}

export interface OrderItem {
  id: number;
  orderId: number;
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  /** @nullable */
  specialInstructions?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  status: OrderStatus;
  total: number;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  /** @nullable */
  stripeStatus?: string | null;
  /** @nullable */
  specialInstructions?: string | null;
  /** @nullable */
  estimatedReadyAt?: string | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderItemBody {
  menuItemId: number;
  quantity: number;
  /** @nullable */
  specialInstructions?: string | null;
}

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  specialInstructions?: string | null;
  items: CreateOrderItemBody[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
  /** @nullable */
  estimatedReadyAt?: string | null;
}

export interface CreatePaymentIntentBody {
  orderId: number;
}

export interface PaymentIntentResponse {
  clientSecret: string;
  paymentIntentId: string;
  amount: number;
}

export interface ConfirmPaymentBody {
  orderId: number;
  paymentIntentId: string;
}

export interface TopItem {
  menuItemId: number;
  name: string;
  totalOrdered: number;
  revenue: number;
}

export interface DashboardStats {
  todayOrders: number;
  todayRevenue: number;
  pendingOrders: number;
  preparingOrders: number;
  completedOrders: number;
  weekRevenue: number;
  monthRevenue: number;
  topItems: TopItem[];
  currentTimingWindow: number;
  isPeakHours: boolean;
}

export interface DailyRevenue {
  date: string;
  revenue: number;
  orders: number;
}

export interface RevenueReport {
  from: string;
  to: string;
  totalRevenue: number;
  totalOrders: number;
  averageOrderValue: number;
  byDay: DailyRevenue[];
}

export interface Expense {
  id: number;
  description: string;
  amount: number;
  category: string;
  /** @nullable */
  vendor?: string | null;
  expenseDate: string;
  isPaid: boolean;
  createdAt: string;
}

export interface CreateExpenseBody {
  description: string;
  amount: number;
  category: string;
  /** @nullable */
  vendor?: string | null;
  expenseDate: string;
  isPaid?: boolean;
}

export interface UpdateExpenseBody {
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  vendor?: string | null;
  /** @nullable */
  expenseDate?: string | null;
  /** @nullable */
  isPaid?: boolean | null;
}

export interface OrderTiming {
  windowMinutes: number;
  isPeakHours: boolean;
  peakStart: string;
  peakEnd: string;
  estimatedReadyAt: string;
}

export interface AiMenuSuggestBody {
  name: string;
  /** @nullable */
  currentDescription?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  price?: number | null;
}

export interface AiMenuSuggestResponse {
  description: string;
  allergenSuggestions: string;
  /** @nullable */
  pricingSuggestion?: string | null;
}

export interface AiSpecialSuggestBody {
  title: string;
  /** @nullable */
  components?: string | null;
}

export interface AiSpecialSuggestResponse {
  description: string;
  /** @nullable */
  suggestedPrice?: number | null;
}

export type ListMenuItemsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  available?: boolean | null;
};

export type ListOrdersParams = {
  /**
   * @nullable
   */
  status?: ListOrdersStatus;
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  offset?: number | null;
};

export type ListOrdersStatus =
  | (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus]
  | null;

export const ListOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetRevenueReportParams = {
  /**
   * @nullable
   */
  from?: string | null;
  /**
   * @nullable
   */
  to?: string | null;
};