# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Skymark Eatery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/categories:
    get:
      operationId: listCategories
      tags: [menu]
      summary: List menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [menu]
      summary: Create a menu category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /menu/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [menu]
      summary: Update a menu category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [menu]
      summary: Delete a menu category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menu/items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: available
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SPECIALS ────────────────────────────────────────────────────────────────
...
  /admin/revenue:
    get:
      operationId: getRevenueReport
      tags: [admin]
      summary: Revenue report (A/R)
      parameters:
        - name: from
          in: query
          schema:
            type: ["string", "null"]
        - name: to
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
  /admin/expenses:
    get:
      operationId: listExpenses
      tags: [admin]
      summary: List expenses (A/P)
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [admin]
      summary: Add an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /admin/expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [admin]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [admin]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        displayOrder:
          type: integer
        createdAt:
          type: string
      required: [id, name, displayOrder, createdAt]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required: [name, displayOrder]
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        displayOrder:
          type: ["integer", "null"]

    # ─── MENU ITEM ─────────────────────────────────────────────────────────────
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        isPopular:
          type: boolean
        allergens:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, price, available, isPopular, createdAt]
    CreateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        isPopular:
          type: boolean
        allergens:
          type: ["string", "null"]
      required: [name, price]
    UpdateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        available:
          type: ["boolean", "null"]
        isPopular:
          type: ["boolean", "null"]
        allergens:
          type: ["string", "null"]

    # ─── SPECIALS ──────────────────────────────────────────────────────────────
...
    RevenueReport:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        averageOrderValue:
          type: number
        byDay:
          type: array
          items:
            $ref: "#/components/schemas/DailyRevenue"
      required: [from, to, totalRevenue, totalOrders, averageOrderValue, byDay]
    DailyRevenue:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required: [date, revenue, orders]
    Expense:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        vendor:
          type: ["string", "null"]
        expenseDate:
          type: string
        isPaid:
          type: boolean
        createdAt:
          type: string
      required: [id, description, amount, category, expenseDate, isPaid, createdAt]
    CreateExpenseBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        vendor:
          type: ["string", "null"]
        expenseDate:
          type: string
        isPaid:
          type: boolean
      required: [description, amount, category, expenseDate]
    UpdateExpenseBody:
      type: object
      properties:
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
        expenseDate:
          type: ["string", "null"]
        isPaid:
          type: ["boolean", "null"]
...