# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skymark Eatery API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemBody {
  /** @nullable */
  categoryId?: number | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  available?: boolean;
  isPopular?: boolean;
  /** @nullable */
  allergens?: string | null;
}