# lib/api-zod/src/generated/types/dailyRevenue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skymark Eatery API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyRevenue {
  date: string;
  revenue: number;
  orders: number;
}