# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skymark Eatery API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  description: string;
  amount: number;
  category: string;
  /** @nullable */
  vendor?: string | null;
  expenseDate: string;
  isPaid: boolean;
  createdAt: string;
}